# backend/openapi.json
{
 "components": {
  "schemas": {
   "AnalyzeBody": {
    "properties": {
     "capture_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Capture Id"
     },
     "device_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Device Id"
     },
     "window_s": {
      "default": 3600,
      "maximum": 604800.0,
      "minimum": 10.0,
      "title": "Window S",
      "type": "integer"
     }
    },
    "title": "AnalyzeBody",
    "type": "object"
   },
...
   "Correlation": {
    "properties": {
     "a_kind": {
      "title": "A Kind",
      "type": "string"
     },
     "a_layer": {
      "enum": [
       "bus",
       "driver",
       "firmware",
       "mac80211",
       "netdev"
      ],
      "title": "A Layer",
      "type": "string"
     },
     "b_kind": {
      "title": "B Kind",
      "type": "string"
     },
     "b_layer": {
      "enum": [
       "bus",
       "driver",
       "firmware",
       "mac80211",
       "netdev"
      ],
      "title": "B Layer",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "incidents_a_before_b": {
      "title": "Incidents A Before B",
      "type": "integer"
     },
     "incidents_with_both": {
      "title": "Incidents With Both",
      "type": "integer"
     },
     "max_window_s": {
      "title": "Max Window S",
      "type": "number"
     },
     "median_lag_s": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Median Lag S"
     },
     "note": {
      "title": "Note",
      "type": "string"
     }
    },
    "required": [
     "id",
     "a_kind",
     "b_kind",
     "a_layer",
     "b_layer",
     "incidents_with_both",
     "incidents_a_before_b",
     "median_lag_s",
     "max_window_s",
     "note"
    ],
    "title": "Correlation",
    "type": "object"
   },
...
   "DiagnosticReport": {
    "properties": {
     "cause_confirmed": {
      "default": false,
      "description": "always false: WHD does not assert root causes",
      "title": "Cause Confirmed",
      "type": "boolean"
     },
     "correlations": {
      "items": {
       "$ref": "#/components/schemas/Correlation"
      },
      "title": "Correlations",
      "type": "array"
     },
     "disclaimer": {
      "title": "Disclaimer",
      "type": "string"
     },
     "follow_ups": {
      "items": {
       "$ref": "#/components/schemas/FollowUp"
      },
      "title": "Follow Ups",
      "type": "array"
     },
     "generated_at": {
      "title": "Generated At",
      "type": "number"
     },
     "hypotheses": {
      "items": {
       "$ref": "#/components/schemas/Hypothesis"
      },
      "title": "Hypotheses",
      "type": "array"
     },
     "incidents": {
      "items": {
       "$ref": "#/components/schemas/Incident"
      },
      "title": "Incidents",
      "type": "array"
     },
     "limits": {
      "items": {
       "type": "string"
      },
      "title": "Limits",
      "type": "array"
     },
     "missing_instrumentation": {
      "items": {
       "$ref": "#/components/schemas/MissingInstrumentation"
      },
      "title": "Missing Instrumentation",
      "type": "array"
     },
     "observations": {
      "items": {
       "$ref": "#/components/schemas/Observation"
      },
      "title": "Observations",
      "type": "array"
     },
     "patterns": {
      "items": {
       "$ref": "#/components/schemas/Pattern"
      },
      "title": "Patterns",
      "type": "array"
     },
     "report_id": {
      "title": "Report Id",
      "type": "string"
     },
     "scope": {
      "$ref": "#/components/schemas/Scope"
     },
     "whd_version": {
      "title": "Whd Version",
      "type": "string"
     }
    },
    "required": [
     "report_id",
     "generated_at",
     "whd_version",
     "scope",
     "cause_confirmed",
     "disclaimer",
     "observations",
     "incidents",
     "patterns",
     "correlations",
     "hypotheses",
     "missing_instrumentation",
     "follow_ups",
     "limits"
    ],
    "title": "DiagnosticReport",
    "type": "object"
   },
...
   "FollowUp": {
    "properties": {
     "distinguishes": {
      "description": "hypothesis ids this test helps separate",
      "items": {
       "type": "string"
      },
      "title": "Distinguishes",
      "type": "array"
     },
     "how": {
      "title": "How",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "safety": {
      "enum": [
       "read-only",
       "needs-user-action"
      ],
      "title": "Safety",
      "type": "string"
     },
     "title": {
      "title": "Title",
      "type": "string"
     },
     "whd_feature": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Whd Feature"
     }
    },
    "required": [
     "id",
     "title",
     "how",
     "distinguishes",
     "safety",
     "whd_feature"
    ],
    "title": "FollowUp",
    "type": "object"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "Hypothesis": {
    "properties": {
     "confidence": {
      "default": "low",
      "enum": [
       "low",
       "moderate"
      ],
      "title": "Confidence",
      "type": "string"
     },
     "contradicting": {
      "items": {
       "type": "string"
      },
      "title": "Contradicting",
      "type": "array"
     },
     "failure_boundary": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "description": "where the evidence places the boundary, if at all",
      "title": "Failure Boundary"
     },
     "follow_up_ids": {
      "items": {
       "type": "string"
      },
      "title": "Follow Up Ids",
      "type": "array"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "missing_instrumentation": {
      "items": {
       "type": "string"
      },
      "title": "Missing Instrumentation",
      "type": "array"
     },
     "statement": {
      "title": "Statement",
      "type": "string"
     },
     "status": {
      "const": "unconfirmed",
      "default": "unconfirmed",
      "title": "Status",
      "type": "string"
     },
     "supporting": {
      "description": "observation ids",
      "items": {
       "type": "string"
      },
      "title": "Supporting",
      "type": "array"
     },
     "supporting_event_ids": {
      "items": {
       "type": "integer"
      },
      "title": "Supporting Event Ids",
      "type": "array"
     },
     "title": {
      "title": "Title",
      "type": "string"
     },
     "unknowns": {
      "items": {
       "type": "string"
      },
      "title": "Unknowns",
      "type": "array"
     }
    },
    "required": [
     "id",
     "title",
     "statement",
     "status",
     "confidence",
     "failure_boundary",
     "supporting",
     "supporting_event_ids",
     "contradicting",
     "unknowns",
     "missing_instrumentation",
     "follow_up_ids"
    ],
    "title": "Hypothesis",
    "type": "object"
   },
   "Incident": {
    "properties": {
     "device_ids": {
      "items": {
       "type": "string"
      },
      "title": "Device Ids",
      "type": "array"
     },
     "duration_s": {
      "title": "Duration S",
      "type": "number"
     },
     "earliest_layers": {
      "description": "layers with a significant event in the first second",
      "items": {
       "enum": [
        "bus",
        "driver",
        "firmware",
        "mac80211",
        "netdev"
       ],
       "type": "string"
      },
      "title": "Earliest Layers",
      "type": "array"
     },
     "end_ns": {
      "title": "End Ns",
      "type": "integer"
     },
     "event_count": {
      "title": "Event Count",
      "type": "integer"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "kinds": {
      "additionalProperties": {
       "type": "integer"
      },
      "title": "Kinds",
      "type": "object"
     },
     "layers": {
      "items": {
       "enum": [
        "bus",
        "driver",
        "firmware",
        "mac80211",
        "netdev"
       ],
       "type": "string"
      },
      "title": "Layers",
      "type": "array"
     },
     "sequence": {
      "description": "first observed significant event per layer, in time order",
      "items": {
       "$ref": "#/components/schemas/IncidentEvent"
      },
      "title": "Sequence",
      "type": "array"
     },
     "severity": {
      "title": "Severity",
      "type": "string"
     },
     "start_ns": {
      "title": "Start Ns",
      "type": "integer"
     }
    },
    "required": [
     "id",
     "device_ids",
     "start_ns",
     "end_ns",
     "duration_s",
     "severity",
     "event_count",
     "layers",
     "sequence",
     "earliest_layers",
     "kinds"
    ],
    "title": "Incident",
    "type": "object"
   },
   "IncidentEvent": {
    "properties": {
     "id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Id"
     },
     "kind": {
      "title": "Kind",
      "type": "string"
     },
     "layer": {
      "enum": [
       "bus",
       "driver",
       "firmware",
       "mac80211",
       "netdev"
      ],
      "title": "Layer",
      "type": "string"
     },
     "rel_s": {
      "title": "Rel S",
      "type": "number"
     },
     "severity": {
      "title": "Severity",
      "type": "string"
     },
     "summary": {
      "title": "Summary",
      "type": "string"
     },
     "ts_boottime_ns": {
      "title": "Ts Boottime Ns",
      "type": "integer"
     }
    },
    "required": [
     "id",
     "ts_boottime_ns",
     "rel_s",
     "layer",
     "severity",
     "kind",
     "summary"
    ],
    "title": "IncidentEvent",
    "type": "object"
   },
...
   "MissingInstrumentation": {
    "properties": {
     "id": {
      "title": "Id",
      "type": "string"
     },
     "proposal": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Proposal"
     },
     "what": {
      "title": "What",
      "type": "string"
     },
     "why_it_matters": {
      "title": "Why It Matters",
      "type": "string"
     }
    },
    "required": [
     "id",
     "what",
     "why_it_matters",
     "proposal"
    ],
    "title": "MissingInstrumentation",
    "type": "object"
   },
...
   "Observation": {
    "description": "A fact derived mechanically from events/inventory. Never interpretive.",
    "properties": {
     "device_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Device Id"
     },
     "evidence_event_ids": {
      "items": {
       "type": "integer"
      },
      "title": "Evidence Event Ids",
      "type": "array"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "kind": {
      "enum": [
       "count",
       "first_last",
       "state",
       "timing",
       "recurrence",
       "coverage"
      ],
      "title": "Kind",
      "type": "string"
     },
     "text": {
      "title": "Text",
      "type": "string"
     }
    },
    "required": [
     "id",
     "kind",
     "text",
     "evidence_event_ids",
     "device_id"
    ],
    "title": "Observation",
    "type": "object"
   },
...
   "Pattern": {
    "properties": {
     "count": {
      "title": "Count",
      "type": "integer"
     },
     "device_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Device Id"
     },
     "evidence_event_ids": {
      "items": {
       "type": "integer"
      },
      "title": "Evidence Event Ids",
      "type": "array"
     },
     "first_ns": {
      "title": "First Ns",
      "type": "integer"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "kind": {
      "title": "Kind",
      "type": "string"
     },
     "last_ns": {
      "title": "Last Ns",
      "type": "integer"
     },
     "max_interval_s": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Interval S"
     },
     "median_interval_s": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Median Interval S"
     },
     "min_interval_s": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Min Interval S"
     },
     "note": {
      "title": "Note",
      "type": "string"
     },
     "regularity": {
      "enum": [
       "periodic",
       "irregular",
       "burst"
      ],
      "title": "Regularity",
      "type": "string"
     }
    },
    "required": [
     "id",
     "device_id",
     "kind",
     "count",
     "first_ns",
     "last_ns",
     "median_interval_s",
     "min_interval_s",
     "max_interval_s",
     "regularity",
     "note",
     "evidence_event_ids"
    ],
    "title": "Pattern",
    "type": "object"
   },
...
   "Scope": {
    "properties": {
     "capture_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Capture Id"
     },
     "demo": {
      "title": "Demo",
      "type": "boolean"
     },
     "device_ids": {
      "items": {
       "type": "string"
      },
      "title": "Device Ids",
      "type": "array"
     },
     "event_count": {
      "title": "Event Count",
      "type": "integer"
     },
     "first_ns": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "First Ns"
     },
     "last_ns": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Ns"
     },
     "significant_event_count": {
      "title": "Significant Event Count",
      "type": "integer"
     },
     "sources": {
      "items": {
       "type": "string"
      },
      "title": "Sources",
      "type": "array"
     },
     "unattributed_events": {
      "default": 0,
      "title": "Unattributed Events",
      "type": "integer"
     },
     "window_s": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Window S"
     }
    },
    "required": [
     "device_ids",
     "window_s",
     "event_count",
     "significant_event_count",
     "sources",
     "first_ns",
     "last_ns",
     "demo",
     "capture_id",
     "unattributed_events"
    ],
    "title": "Scope",
    "type": "object"
   },
...
   "SummarizeBody": {
    "properties": {
     "model": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Model"
     },
     "redact_macs": {
      "default": true,
      "title": "Redact Macs",
      "type": "boolean"
     },
     "report_id": {
      "title": "Report Id",
      "type": "string"
     }
    },
    "required": [
     "report_id"
    ],
    "title": "SummarizeBody",
    "type": "object"
   },
   "Summary": {
    "properties": {
     "elapsed_s": {
      "title": "Elapsed S",
      "type": "number"
     },
     "eval_tokens": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Eval Tokens"
     },
     "generated_by": {
      "title": "Generated By",
      "type": "string"
     },
     "model": {
      "title": "Model",
      "type": "string"
     },
     "not_evidence": {
      "default": true,
      "title": "Not Evidence",
      "type": "boolean"
     },
     "prompt_chars": {
      "title": "Prompt Chars",
      "type": "integer"
     },
     "report_id": {
      "title": "Report Id",
      "type": "string"
     },
     "text": {
      "title": "Text",
      "type": "string"
     },
     "warnings": {
      "items": {
       "type": "string"
      },
      "title": "Warnings",
      "type": "array"
     }
    },
    "required": [
     "generated_by",
     "model",
     "report_id",
     "text",
     "elapsed_s",
     "eval_tokens",
     "warnings",
     "not_evidence",
     "prompt_chars"
    ],
    "title": "Summary",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "Wireless Hardware Debugger",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/diagnostics/analyze": {
   "post": {
    "operationId": "analyze_api_v1_diagnostics_analyze_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AnalyzeBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DiagnosticReport"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Analyze"
   }
  },
...
  "/api/v1/diagnostics/llm/summarize": {
   "post": {
    "operationId": "llm_summarize_api_v1_diagnostics_llm_summarize_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SummarizeBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Summary"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Llm Summarize"
   }
  },
  "/api/v1/diagnostics/reports/{report_id}": {
   "get": {
    "operationId": "get_report_api_v1_diagnostics_reports__report_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "report_id",
      "required": true,
      "schema": {
       "title": "Report Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DiagnosticReport"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Report"
   }
  },
...
 }
}